handle decompression failure during transition from 1 to 2 IP headers
The transition between 1 to 2 IP headers (or the reverse transition) was
already handled by the ROHC decompressor since ages. However, a special
case was not.

If the decompressor fails to decompress the packet that changes the
number of IP headers, the decompression context is updated any way and
the decompression of the next packet may fail because of this bad
update.

This change defers the context update until the decompression is deemed
successful.

The problem was detected with AFL. A robustness test was added.