RTP profile: accept 0 bit of SN for IR w/o dynamic chain
The ROHC compressor does not build nor send IR packets without a dynamic
chain, although it is optional for the RTP profile. This leads to the
ROHC decompressor being under-tested for this kind of packets.

This change makes the decompressor accept to get 0 bit of SN retrieved
from a ROHC packet since it is the case for IR packets w/o dynamic
chain.

The problem was detected by AFL. Several robustness tests were added.
One of the tests doesn't fully succeed currently since another problem
related to the UDP checksum occurs.