RFC3095: only RTP profile uses the x bit in the packet type byte
The packet type byte in the RFC3095 profiles contains one x bit that is
profile-specific. All of the profiles except the RTP profile define this
x field as reserved (and set it to value 0). Only the RTP profile
defines this bit as the D flag to specify whether the dynamic chain is
present or not in the IR packet.

This change forces the x bit to 0 for all RFC3095 profiles except the
RTP profile. Many reference captures for the non-regression tests
therefore change.