RFC3095: better handle transition between IP versions
An assertion occured if a ROHC packet changed the version of the outer
or inner IP header w/o bits for the v4- or v6-specific fields (ie. DF,
NBO, RND, SID, IP-ID, Flow Label). The decompressor tried to retrieve
the information from an uninitialized part of the context.

This change adds checks for the 4-to-6 or 6-to-4 switch before using
the context information. If not initialized, decompression fails.

Detected with AFL. A robustness test was added.