RTP profile: soft reject of packets with no UDP checksum bits
The ROHC compressor does not build nor send IR packets without a dynamic
chain, although it is optional for the RTP profile. This leads to the
ROHC decompressor being under-tested for this kind of packets.

This change makes the decompressor rejects the ROHC packets that
transmit no bit of UDP checksum while the UDP checksum is not omitted.
This may be the case with IR packets w/o dynamic chain.

The problem was detected by AFL. A robustness test was added.