RTP profile: handle short extension 3 for UOR-2-TS packet
A heap overflow occured if the decompressor tried to decompress a ROHC
UOR-2-TS packet with a too short extension 3. The packet shall be long
enough to contain the 1st byte of flags of the extension 3, but short
enough not to contain the 2nd byte with the optional ip2 flag.

This change correctly checks the length of the ROHC packet before
parsing the 2nd byte of flag.

Detected with AFL. A robustness test was added.