RFC3095: extension 3: avoid sending inner IP header flags
For RTP profiles, the inner IP header flags contains the ip2 flag that
specifies whether the outer IP header flags are present or not. So, the
inner IP header flags shall be present whenever the outer IP header
flags are present.

The ROHC library was checking for such a situation, but the check was
wrong. The check tested the RTP profile and the presence of a 2nd IP
header, but not the value of the ip2 flag itself.

This resulted in ROHC headers with 1 superflous byte for the inner IP
header flags when:
 * the compressed stream is one RTP stream,
 * the compressed stream contains 2 IP headers,
 * no IP field changed in the inner IP header,
 * no IP field changed in the outer IP header.